# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/kitchens/{kitchenId}": {
      "get": {
        "operationId": "KitchensController_getOne",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KitchenDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Szczegóły kuchni i członkowie",
        "tags": [
          "kitchens"
        ]
      },
      "delete": {
        "operationId": "KitchensController_remove",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Usunięcie kuchni przez właściciela (kaskadowo członkowie, zaproszenia, produkty i partie)",
        "tags": [
          "kitchens"
        ]
      }
    },
...
    "/api/kitchens/{kitchenId}/media/uploads": {
      "post": {
        "operationId": "MediaController_beginUpload",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeginMediaUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeginMediaUploadResultDto"
                }
              }
            }
          }
        },
        "summary": "Rozpoczęcie wysyłki zdjęcia",
        "tags": [
          "media"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/media/{mediaAssetId}/memory-upload": {
      "post": {
        "operationId": "MediaController_memoryUpload",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaAssetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryUploadDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Wysyłka zawartości do magazynu w pamięci (tylko MEDIA_STORAGE_DRIVER=memory)",
        "tags": [
          "media"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/media/{mediaAssetId}/complete": {
      "post": {
        "operationId": "MediaController_completeUpload",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaAssetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAssetDto"
                }
              }
            }
          }
        },
        "summary": "Zakończenie wysyłki i przetworzenie zdjęcia",
        "tags": [
          "media"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/media/{mediaAssetId}": {
      "delete": {
        "operationId": "MediaController_deleteAsset",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaAssetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Usunięcie zdjęcia wraz z plikami",
        "tags": [
          "media"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/products/{productId}/image": {
      "post": {
        "operationId": "MediaAttachmentController_attachProductImage",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachedMediaDto"
                }
              }
            }
          }
        },
        "summary": "Przypisanie zdjęcia do produktu",
        "tags": [
          "media"
        ]
      },
      "delete": {
        "operationId": "MediaAttachmentController_detachProductImage",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Odpięcie zdjęcia produktu",
        "tags": [
          "media"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/recipes/{recipeId}/cover": {
      "post": {
        "operationId": "MediaAttachmentController_attachRecipeCover",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recipeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachedMediaDto"
                }
              }
            }
          }
        },
        "summary": "Przypisanie okładki przepisu (autor)",
        "tags": [
          "media"
        ]
      },
      "delete": {
        "operationId": "MediaAttachmentController_detachRecipeCover",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recipeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Odpięcie okładki przepisu (autor)",
        "tags": [
          "media"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/recipes/{recipeId}/steps/{stepId}/image": {
      "post": {
        "operationId": "MediaAttachmentController_attachRecipeStepImage",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recipeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stepId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachedMediaDto"
                }
              }
            }
          }
        },
        "summary": "Przypisanie zdjęcia do kroku przepisu (autor)",
        "tags": [
          "media"
        ]
      },
      "delete": {
        "operationId": "MediaAttachmentController_detachRecipeStepImage",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recipeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stepId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Odpięcie zdjęcia kroku przepisu (autor)",
        "tags": [
          "media"
        ]
      }
    },
...
  },
  "info": {
    "title": "Moja Kuchnia API",
    "description": "REST API aplikacji Moja Kuchnia",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
...
      "KitchenMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "member"
            ]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "userId",
          "email",
          "name",
          "role",
          "joinedAt"
        ]
      },
      "KitchenDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KitchenMemberDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "members"
        ]
      },
...
      "MediaUploadTargetDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "recipeId": {
            "type": "string",
            "format": "uuid"
          },
          "recipeStepId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "BeginMediaUploadDto": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "product",
              "recipe_cover",
              "recipe_step"
            ]
          },
          "declaredMimeType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "example": "image/jpeg",
            "description": "Deklarowany typ pliku. Zawartość jest weryfikowana po bajtach."
          },
          "declaredByteSize": {
            "type": "number",
            "example": 512000
          },
          "target": {
            "$ref": "#/components/schemas/MediaUploadTargetDto"
          }
        },
        "required": [
          "purpose",
          "declaredMimeType",
          "declaredByteSize"
        ]
      },
      "BeginMediaUploadHeadersDto": {
        "type": "object",
        "properties": {
          "Content-Type": {
            "type": "string",
            "example": "image/jpeg"
          }
        },
        "required": [
          "Content-Type"
        ]
      },
      "BeginMediaUploadResultDto": {
        "type": "object",
        "properties": {
          "mediaAssetId": {
            "type": "string"
          },
          "uploadUrl": {
            "type": "string",
            "description": "Adres do wysłania pliku metodą PUT. Dla sterownika memory jest to endpoint API."
          },
          "objectKey": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "headers": {
            "$ref": "#/components/schemas/BeginMediaUploadHeadersDto"
          }
        },
        "required": [
          "mediaAssetId",
          "uploadUrl",
          "objectKey",
          "expiresAt",
          "headers"
        ]
      },
      "MemoryUploadDto": {
        "type": "object",
        "properties": {
          "contentBase64": {
            "type": "string",
            "description": "Zawartość pliku zakodowana w base64."
          }
        },
        "required": [
          "contentBase64"
        ]
      },
      "MediaImageDto": {
        "type": "object",
        "properties": {
          "mediaAssetId": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "description": "Krótko żyjący podpisany URL. Nie zapisuj go."
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "mediaAssetId",
          "url",
          "thumbnailUrl"
        ]
      },
      "MediaAssetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kitchenId": {
            "type": "string"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "product",
              "recipe_cover",
              "recipe_step"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "ready",
              "failed"
            ]
          },
          "mimeType": {
            "type": "string"
          },
          "byteSize": {
            "type": "number"
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          },
          "image": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaImageDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "kitchenId",
          "purpose",
          "status",
          "mimeType",
          "byteSize",
          "width",
          "height",
          "image",
          "createdAt",
          "updatedAt"
        ]
      },
      "AttachMediaDto": {
        "type": "object",
        "properties": {
          "mediaAssetId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "mediaAssetId"
        ]
      },
      "AttachedMediaDto": {
        "type": "object",
        "properties": {
          "targetId": {
            "type": "string",
            "description": "Id produktu, przepisu albo kroku przepisu."
          },
          "image": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaImageDto"
              }
            ]
          }
        },
        "required": [
          "targetId",
          "image"
        ]
      },
...
    }
  }
}